requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.36
psycopg2-binary>=2.9.9
pydantic>=2.7.0
python-dotenv>=1.0.0
reportlab>=4.2.0
requests>=2.32.0
pytest>=8.2.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
jinja2>=3.1.0
python-multipart>=0.0.9